+++ pkg/alertmanager/alertmanager_http_test.go
@@ -37,6 +37,7 @@ func TestMultitenantAlertmanager_GetStatusHandler(t *testing.T) {
 				cluster.DefaultTcpTimeout,
 				cluster.DefaultProbeTimeout,
 				cluster.DefaultProbeInterval,
+				nil,
 			)
 		}