+++ pkg/querier/querier.go
@@ -268,7 +268,8 @@ func (q querier) Select(_ bool, sp *storage.SelectHints, matchers ...*labels.Mat
 	defer log.Span.Finish()
 
 	if sp != nil {
-		level.Debug(log).Log("start", util.TimeFromMillis(sp.Start).UTC().String(), "end", util.TimeFromMillis(sp.End).UTC().String(), "step", sp.Step, "matchers", matchers)
+		level.Debug(log).Log("start", util.TimeFromMillis(sp.Start).UTC().String(), "end",
+			util.TimeFromMillis(sp.End).UTC().String(), "step", sp.Step, "matchers", util.MatchersStringer(matchers))
 	}
 
 	// Kludge: Prometheus passes nil SelectHints if it is doing a 'series' operation,