+++ pkg/querier/queryrange/roundtrip.go
@@ -47,7 +47,6 @@ type Config struct {
 	CacheResults           bool `yaml:"cache_results"`
 	MaxRetries             int  `yaml:"max_retries"`
 	ShardedQueries         bool `yaml:"parallelise_shardable_queries"`
-	TotalShards            int  `yaml:"total_shards"`
 }
 
 // RegisterFlags adds the flags required to config this to the given FlagSet.
@@ -57,7 +56,6 @@ func (cfg *Config) RegisterFlags(f *flag.FlagSet) {
 	f.BoolVar(&cfg.AlignQueriesWithStep, "querier.align-querier-with-step", false, "Mutate incoming queries to align their start and end with their step.")
 	f.BoolVar(&cfg.CacheResults, "querier.cache-results", false, "Cache query results.")
 	f.BoolVar(&cfg.ShardedQueries, "querier.parallelise-shardable-queries", false, "Perform query parallelisations based on storage sharding configuration and query ASTs. This feature is supported only by the blocks storage engine.")
-	f.IntVar(&cfg.TotalShards, "querier.total-shards", 16, "The amount of shards to use when doing parallelisation via query sharding by default.")
 	cfg.ResultsCacheConfig.RegisterFlags(f)
 }
 
@@ -70,9 +68,6 @@ func (cfg *Config) Validate() error {
 		if err := cfg.ResultsCacheConfig.Validate(); err != nil {
 			return errors.Wrap(err, "invalid ResultsCache config")
 		}
-	}
-	if cfg.ShardedQueries && cfg.TotalShards <= 0 {
-		return errors.New("querier.total-shards must be > 0 when parallelisation of shardable queries is enabled")
 	}
 	return nil
 }
@@ -165,7 +160,7 @@ func NewTripperware(
 	var c cache.Cache
 	if cfg.CacheResults {
 		shouldCache := func(r Request) bool {
-			return !r.GetCachingOptions().Disabled
+			return !r.GetOptions().CacheDisabled
 		}
 		queryCacheMiddleware, cache, err := NewResultsCacheMiddleware(log, cfg.ResultsCacheConfig, constSplitter(cfg.SplitQueriesByInterval), limits, codec, cacheExtractor, cacheGenNumberLoader, shouldCache, registerer)
 		if err != nil {
@@ -189,7 +184,7 @@ func NewTripperware(
 			NewQueryShardingMiddleware(
 				log,
 				promql.NewEngine(engineOpts),
-				cfg.TotalShards,
+				limits,
 				registerer,
 			),
 		)